% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cars}
\alias{cars}
\title{Cars Dataset}
\format{
A data frame with 50 observations on 2 variables:
\describe{
\item{speed}{A numeric vector representing the speed of the car in mph.}
\item{dist}{A numeric vector representing the stopping distance of the car in feet.}
}
}
\source{
The built-in R dataset \code{cars}.
}
\usage{
cars
}
\description{
This dataset contains speed and stopping distances of cars.
}
\keyword{datasets}
