% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{mymean}
\alias{mymean}
\title{Calculate Mean of a Numeric Vector}
\usage{
mymean(x)
}
\arguments{
\item{x}{A numeric vector for which the mean is computed.}
}
\value{
A numeric value representing the mean of x.
}
\description{
This function returns the mean of the given numeric vector.
}
